+++ inst/doc/An_Introduction_to_packager.R
@@ -2,26 +2,32 @@
 path <- file.path(tempdir(), "myFirstPackage")
 packager::create(path, fakemake = "check")
 
+
 ## -----------------------------------------------------------------------------
 list.files(path, recursive = FALSE)
 r <- git2r::repository(path)
 summary(r)
 git2r::status(r)
 
+
 ## -----------------------------------------------------------------------------
 cat(readLines(file.path(path, "log", "spell.Rout")), sep = "\n")
 tail(readLines(file.path(path, "log", "check.Rout")), sep = "\n")
 
+
 ## -----------------------------------------------------------------------------
 cat(readLines(file.path(path, "TODO.md")), sep = "\n")
 
+
 ## -----------------------------------------------------------------------------
 cat(readLines(file.path(path, "DESCRIPTION")), sep = "\n")
+
 
 ## -----------------------------------------------------------------------------
 unlink(path, recursive = TRUE)
 if ("myFirstPackage" %in% .packages()) detach("package:myFirstPackage", 
                                               unload = TRUE)
+
 
 ## ---- results = "hide", message = FALSE---------------------------------------
 package_title <- "veryImportantPackage"
@@ -35,55 +41,65 @@ packager::create(path, author_at_r = a, title = package_title,
 ## -----------------------------------------------------------------------------
 cat(readLines(file.path(path, "DESCRIPTION")), sep = "\n")
 
+
 ## ---- eval = FALSE------------------------------------------------------------
-#  pkgload::load_all(path)
-#  help(paste0(package_title, "-package"))
+## pkgload::load_all(path)
+## help(paste0(package_title, "-package"))
 
 ## ---- echo = FALSE------------------------------------------------------------
 pkgload::load_all(path)
-# insert developement page
 help_file <-  system.file("man", paste0(package_title, "-package.Rd"), 
                           package = devtools::as.package(path)$package)
 captured <- gsub('_\b', '',  capture.output(tools:::Rd2txt(help_file) ))
 cat(captured, sep = "\n")
+
 
 ## -----------------------------------------------------------------------------
 r <- git2r::repository(path)
 summary(r)
 git2r::status(r)
 
+
 ## -----------------------------------------------------------------------------
 list.files(file.path(path, "log"))
+
 
 ## -----------------------------------------------------------------------------
 ml <- packager::get_package_makelist(is_cran = TRUE)
 cbind(lapply(ml, function(x) x[["target"]]),
       lapply(ml, function(x) x[["alias"]]))
 
+
 ## -----------------------------------------------------------------------------
 suppressMessages(withr::with_dir(path, 
                                   print(fakemake::make("build", ml, 
                                                        verbose = FALSE))))
 
+
 ## -----------------------------------------------------------------------------
 cat(git2r::diff(r, as_char = TRUE, path = file.path(".Rbuildignore")))
+
 
 ## -----------------------------------------------------------------------------
 git2r::status(r)
 cat(diff(r, as_char = TRUE, path = ".Rbuildignore"))
+
 
 ## -----------------------------------------------------------------------------
 withr::with_dir(path, packager::git_add_commit(path = ".", untracked = TRUE,
                                                message = "make build"))
 git2r::status(r)
 
+
 ## -----------------------------------------------------------------------------
 suppressMessages(withr::with_dir(path, 
                                  print(fakemake::make("check", ml, 
                                                       verbose = FALSE))))
 
+
 ## -----------------------------------------------------------------------------
 git2r::status(r)
+
 
 ## -----------------------------------------------------------------------------
 cat(tail(readLines(file.path(path, "log", "check.Rout")), n = 7), sep = "\n")
@@ -91,19 +107,24 @@ check_log <- file.path(path, "log", "check.Rout")
 status <- packager::get_check_status(check_log)
 RUnit::checkEqualsNumeric(status[["status"]][["errors"]], 0)
 
+
 ## -----------------------------------------------------------------------------
 withr::with_dir(path, packager::git_add_commit(path = ".", untracked = TRUE,
                                                message = "make check"))
 
+
 ## -----------------------------------------------------------------------------
 system.time(withr::with_dir(path, print(fakemake::make("check", ml, verbose = FALSE))))
+
 
 ## -----------------------------------------------------------------------------
 withr::with_dir(path, print(fakemake::make("cran_comments", ml, verbose = FALSE)))
 cat(readLines(file.path(path, "cran-comments.md")), sep = "\n")
 
+
 ## -----------------------------------------------------------------------------
 try(packager::submit(path))
+
 
 ## -----------------------------------------------------------------------------
 packager::git_add_commit(path = path, untracked = TRUE, 
@@ -111,8 +132,10 @@ packager::git_add_commit(path = path, untracked = TRUE,
 git2r::status(r)
 
 
+
 ## -----------------------------------------------------------------------------
 try(packager::submit(path))
+
 
 ## -----------------------------------------------------------------------------
 packager::git_tag(path = path, message = "A Tag")