+++ inst/runit_tests/test_main.R
@@ -16,10 +16,31 @@ test_create <- function() {
                   "NEWS.md", "R/prutp-package.R", "R/throw.R", "README.Rmd",
                   "tests/runit.R", "tests/testthat.R",
                   "tests/testthat/test-throw.R",
-                  "vignettes/An_Introduction_to_prutp.Rmd")
+                  "vignettes/An_Introduction_to_prutp.Rasciidoc")
     RUnit::checkTrue(all(expected %in% result))
     # FIXME: I cannot test on fakemake = "check" as this fails with strange
     # conditions. So I ran it 
     # in create()'s examples, now it is located in test/testthat/test-main.R
 }
-if (interactive()) test_create() 
+test_create_rmd <- function() {
+    old_opts <- options(warn = 1)
+    on.exit(old_opts)
+    d <- file.path(tempdir(), "prutp")
+    on.exit(unlink(d, recursive = TRUE))
+    # use the classic Rmd-vignette
+    packager::create(path = d, fakemake = "roxygen2", 
+                     use_rasciidoc_vignette = FALSE)
+    files <- sort(dir(d, recursive = TRUE, full.names = FALSE))
+    # checking on file contents does not work as covr and RUnit give
+    # different digest::sha1()-values.
+    result <- files
+    expected <- c("DESCRIPTION", "devel.R",
+                  "inst/runit_tests/runit-throw.R", "LICENSE", "Makefile",
+                  "man/prutp-package.Rd", "man/throw.Rd", "NAMESPACE",
+                  "NEWS.md", "R/prutp-package.R", "R/throw.R", "README.Rmd",
+                  "tests/runit.R", "tests/testthat.R",
+                  "tests/testthat/test-throw.R",
+                  "vignettes/An_Introduction_to_prutp.Rmd")
+    RUnit::checkTrue(all(expected %in% result))
+}
+if (interactive()) test_create_rmd()